+++ demo-source/package.json
@@ -14,8 +14,8 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "npx webpack --config service-worker-webpack.config.ts; react-scripts start",
-    "build": "npx webpack --config service-worker-webpack.config.ts; react-scripts build; cat aphex.txt >> build/index.html; cp build/index.html build/404.html; git add docs/static/js/*",
+    "start": "npx webpack --config service-worker-dev.config.ts  && react-scripts start",
+    "build": "npx webpack --config service-worker-prod.config.ts && react-scripts build; cat aphex.txt >> build/index.html && cp build/index.html build/404.html && git add docs/static/js/* && rm demo-source/public/*.map",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
