+++ demo-source/src/Home.css
@@ -13,7 +13,7 @@
 .home h1 {
 	margin: 0;
 	padding: 0;
-	margin-bottom: 0.25rem;
+	margin-bottom: 0.75rem;
 }
 
 .home-menu .big-link:not(:last-child) {
