+++ docs/asset-manifest.json
@@ -1,7 +1,7 @@
 {
   "files": {
-    "main.css": "/php-wasm/static/css/main.b43dd4ba.css",
-    "main.js": "/php-wasm/static/js/main.ef0052d5.js",
+    "main.css": "/php-wasm/static/css/main.8f830611.css",
+    "main.js": "/php-wasm/static/js/main.df161e95.js",
     "static/media/php-web.mjs.wasm": "/php-wasm/static/media/php-web.mjs.855228a71a5b71165d46.wasm",
     "worker-xquery.js": "/php-wasm/16470d85f9e6b184411b622ce1783fd0.js",
     "worker-javascript.js": "/php-wasm/70896c45cb42a1a16168cef789116031.js",
@@ -491,11 +491,11 @@
     "ftl.js": "/php-wasm/f76ee9c8abfdd96fb9d70116d40435d5.js",
     "d.js": "/php-wasm/c98e74fc97b04fe8bf43dcdff549afcf.js",
     "mode-text.js": "/php-wasm/0ffb18fb70c87335edee31a479f58a43.js",
-    "main.b43dd4ba.css.map": "/php-wasm/static/css/main.b43dd4ba.css.map",
-    "main.ef0052d5.js.map": "/php-wasm/static/js/main.ef0052d5.js.map"
+    "main.8f830611.css.map": "/php-wasm/static/css/main.8f830611.css.map",
+    "main.df161e95.js.map": "/php-wasm/static/js/main.df161e95.js.map"
   },
   "entrypoints": [
-    "static/css/main.b43dd4ba.css",
-    "static/js/main.ef0052d5.js"
+    "static/css/main.8f830611.css",
+    "static/js/main.df161e95.js"
   ]
 }
